Improve GTK App
