+++ gtk/src/app/content/error.rs
