+++ gtk/src/app/content/summary.rs
