--- gtk/src/ui/ui.css
@@ -1,68 +1,0 @@
-.left-panel {
-    margin-right: .5em;
-}
-
-.left-panel image {
-    padding: .25em;
-}
-
-.right-panel {
-    padding: .5em;
-}
-
-spinner {
-    padding-left: .5em;
-}
-
-.h2 {
-	font-size: 1.25em;
-	font-weight: bold;
-	padding-bottom: 1em;
-}
-
-.bold {
-    font-weight: bold;    
-}
-
-.desc {
-    padding-bottom: 1em;
-}
-
-.hash-box button, .hash-box entry {
-    border-radius: 0;
-}
-
-.devices {
-    border-width:0.2em;
-    border-style: solid;
-    border-color: @theme_fg_color;
-}
-
-.devices > row:nth-child(1) {
-    border-bottom-width:0.2em;
-    border-style: solid;
-    border-color: @theme_fg_color;
-    font-weight: bold;
-}
-
-.progress-label {
-    font-weight: bold;
-    padding-right: 1em;
-}
-
-grid {
-    margin-top: 1em;
-    padding-top: .5em;
-    border-top-width: .2em;
-    border-top-style: solid;
-}
-
-box > spinner {
-    padding-top: .25em;
-    padding-right: .5em;
-}
-
-grid box, grid > label {
-    padding-top: .5em;
-    padding-bottom: .5em;
-}
